fix: the todo items view now has its own separate component
